+++ src/main/java/mackeabit/shop/vo/OrdersVO.java
@@ -7,7 +7,7 @@ import lombok.Setter;
 @Setter
 public class OrdersVO {
 
-    private Integer order_idx, member_idx, pd_idx, order_mi,
+    private Long order_idx, member_idx, pd_idx, order_mi,
             now_price;
     private String address, order_date;
